api/requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
asgiref
gunicorn
python-dotenv
pytz
sqlparse
flake8
factory_boy
coverage
